[Wizard] split test file into two smaller
